test suite added to data-stringmap
