+++ Data/StringMap/FuzzySearch.hs
@@ -12,45 +12,43 @@
   Portability: not portable
 
   Functions for fuzzy search in a prefix tree
-  
+
 -}
 
 -- ----------------------------------------------------------------------------
 
 module Data.StringMap.FuzzySearch
-	(
-	prefixFindCaseWithKey
-	, prefixFindNoCaseWithKey
-	, prefixFindNoCase
-	, lookupNoCase
-	, lookupNoCaseBF --redundant
-	, prefixFindCaseWithKeyBF
-	, prefixFindNoCaseWithKeyBF
-	, noCaseKeys
-	, noLowerCaseKeys
-	, noCasePS
-	, noLowerCasePS
-	, noUmlautPS
-)
+    ( prefixFindCaseWithKey
+    , prefixFindNoCaseWithKey
+    , prefixFindNoCase
+    , lookupNoCase
+    , lookupNoCaseBF --redundant
+    , prefixFindCaseWithKeyBF
+    , prefixFindNoCaseWithKeyBF
+    , noCaseKeys
+    , noLowerCaseKeys
+    , noCasePS
+    , noLowerCasePS
+    , noUmlautPS
+    )
 where
 
 import           Data.Char
 
 import           Data.StringMap.Base
 import           Data.StringMap.StringSet
---import           Data.StringMap.Types
 
 -- ----------------------------------------
 
 -- | /O(max(L,R))/ Find all values where the string is a prefix of the key.
 
-prefixFindCaseWithKey           :: Key -> StringMap a -> [(Key, a)] 
-prefixFindCaseWithKey k         = toList . cutPx' (singlePS k) 
+prefixFindCaseWithKey           :: Key -> StringMap a -> [(Key, a)]
+prefixFindCaseWithKey k         = toList . cutPx' (singlePS k)
 
-prefixFindNoCaseWithKey         :: Key -> StringMap a -> [(Key, a)] 
-prefixFindNoCaseWithKey k       = toList . cutPx' (noCaseKeys k) 
+prefixFindNoCaseWithKey         :: Key -> StringMap a -> [(Key, a)]
+prefixFindNoCaseWithKey k       = toList . cutPx' (noCaseKeys k)
 
-prefixFindNoCase                :: Key -> StringMap a -> [a] 
+prefixFindNoCase                :: Key -> StringMap a -> [a]
 prefixFindNoCase k              = elems . cutPx' (noCaseKeys k)
 
 lookupNoCase                    :: Key -> StringMap a -> [(Key, a)]
@@ -61,11 +59,11 @@ lookupNoCase k                  = toList . cutAllPx' (noCaseKeys k)
 -- | /O(max(L,R))/ Find all values where the string is a prefix of the key.
 -- Breadth first variant, short words first in the result list
 
-prefixFindCaseWithKeyBF         :: Key -> StringMap a -> [(Key, a)] 
-prefixFindCaseWithKeyBF k       = toListBF . cutPx' (singlePS k) 
+prefixFindCaseWithKeyBF         :: Key -> StringMap a -> [(Key, a)]
+prefixFindCaseWithKeyBF k       = toListBF . cutPx' (singlePS k)
 
-prefixFindNoCaseWithKeyBF       :: Key -> StringMap a -> [(Key, a)] 
-prefixFindNoCaseWithKeyBF k     = toListBF . cutPx' (noCaseKeys k) 
+prefixFindNoCaseWithKeyBF       :: Key -> StringMap a -> [(Key, a)]
+prefixFindNoCaseWithKeyBF k     = toListBF . cutPx' (noCaseKeys k)
 
 lookupNoCaseBF                  :: Key -> StringMap a -> [(Key, a)]
 lookupNoCaseBF k                = toListBF . cutAllPx' (noCaseKeys k)
@@ -83,7 +81,7 @@ noLowerCaseKeys         = noLowerCasePS . singlePS
 
 noCasePS                        :: StringSet -> StringSet
 noCasePS                        = fuzzyCharPS (\ x -> [toUpper x, toLower x])
-                                      
+
 noLowerCasePS                   :: StringSet -> StringSet
 noLowerCasePS                   = fuzzyCharPS (\ x -> [toUpper x, x])
