+++ data-stringmap.cabal
@@ -1,18 +1,20 @@
-name: data-stringmap
-version: 1.0
-license: MIT
+name:         data-stringmap
+version:      1.0
+license:      MIT
 license-file: LICENSE
-author: Uwe Schmidt
-maintainer: uwe@fh-wedel.de
-bug-reports: https://github.com/sebastian-philipp/StringMap/issues
-synopsis: An efficient implementation of maps from strings to arbitrary values
-category: Data Structures
-description:
-    An efficient implementation of maps from strings to arbitrary values.
-    Values can associated with an arbitrary byte key. Searching for keys is very fast, but
-    the prefix tree probably consumes more memory than "Data.Map". The main differences are the special
-    'prefixFind' functions, which can be used to perform prefix queries. The interface is
-build-type: Simple
+author:       SebastianPhilipp, Uwe Schmidt
+maintainer:   uwe@fh-wedel.de
+bug-reports:  https://github.com/sebastian-philipp/StringMap/issues
+synopsis:     An efficient implementation of maps from strings to arbitrary values
+category:     Data Structures
+description:  An efficient implementation of maps from strings to arbitrary values.
+              Values can associated with an arbitrary byte key.
+              Searching for keys is very fast, but
+              the prefix tree probably consumes more memory than
+              "Data.Map". The main differences are the special
+              'prefixFind' functions, which can be used to perform prefix queries.
+
+build-type:   Simple
 cabal-version:  >=1.8
 extra-source-files:
     tests/*.hs
@@ -23,9 +25,12 @@ source-repository head
     location: https://github.com/sebastian-philipp/StringMap.git
 
 Library
-    build-depends: base >= 4.5 && < 5, deepseq >= 1.2, binary >= 0.5, containers >= 0.4
+    build-depends: base       >= 4.5,
+                   deepseq    >= 1.2,
+                   binary     >= 0.5,
+                   containers >= 0.4
 
-    ghc-options: -O2 -Wall
+    ghc-options: -O2 -Wall -fwarn-tabs -fno-warn-unused-binds
 
     exposed-modules:
         Data.StringMap
@@ -37,3 +42,18 @@ Library
     other-modules:
         Data.StringMap.Base
         Data.StringMap.FuzzySearch
+
+test-suite properties
+  type:           exitcode-stdio-1.0
+  main-is:        StringMapProperties.hs
+  build-depends:  data-stringmap,
+                  base                       >= 4.5,
+                  containers                 >= 0.4,
+                  HUnit                      >= 1.2,
+                  QuickCheck                 >= 2.4,
+                  test-framework             >= 0.6,
+                  test-framework-quickcheck2 >= 0.2,
+                  test-framework-hunit       >= 0.2
+
+  ghc-options:    -Wall -fwarn-tabs
+  hs-source-dirs: tests
